added props and began adding box colliders to the corridor walls
I had to delete the Library folder again before I could commit to main. Apparently that's what you're supposed to do when sharing unity projects:

"All the data for Unity’s internal representation of your assets is stored in the Library folder which can be thought of as similar to a cache folder. As a user, you should never have to alter the Library folder manually and attempting to do so may affect the functioning of the project in the Unity editor. However, it is always safe to delete the Library folder (while the project is not open in Unity) as all its data is generated from what is stored in the Assets and ProjectSettings folders. This also means that the Library folder should not be included in version control." https://docs.unity3d.com/560/Documentation/Manual/BehindtheScenes.html